Move ChunkInsert to new file
